move rename to seperate feature
