+++ starsky/starsky.foundation.database/Interfaces/IQuery.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using starsky.foundation.database.Models;
@@ -64,6 +65,7 @@ namespace starsky.foundation.database.Interfaces
         FileIndexItem UpdateItem(FileIndexItem updateStatusContent);
         List<FileIndexItem> UpdateItem(List<FileIndexItem> updateStatusContentList);
 
+        [Obsolete("use PathHelper.RemoveLatestSlash()")]
         string SubPathSlashRemove(string subPath = "/");
 
         RelativeObjects GetNextPrevInFolder(string currentFolder);
@@ -71,6 +73,6 @@ namespace starsky.foundation.database.Interfaces
         List<FileIndexItem> StackCollections(List<FileIndexItem> databaseSubFolderList);
         void CacheUpdateItem(List<FileIndexItem> updateStatusContent);
 
-	    bool IsCacheEnabled();
+        Task AddParentItemsAsync(string subPath);
     }
 }
