+++ starsky/starsky.sln
@@ -69,6 +69,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.metaupdate"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.export", "starsky.feature.export\starsky.feature.export.csproj", "{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "starsky.feature.rename", "starsky.feature.rename\starsky.feature.rename.csproj", "{D59FC5C7-11E8-4BE6-A500-74088D74F841}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -167,6 +169,10 @@ Global
 		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D59FC5C7-11E8-4BE6-A500-74088D74F841}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D59FC5C7-11E8-4BE6-A500-74088D74F841}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D59FC5C7-11E8-4BE6-A500-74088D74F841}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D59FC5C7-11E8-4BE6-A500-74088D74F841}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -197,5 +203,6 @@ Global
 		{0F47FD1D-CE79-4069-AFD4-252ED9D9445E} = {C757587F-8490-4815-97CE-9BE39A6C82C6}
 		{FBB448C3-A995-469C-8004-11FA006F65C5} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 		{A0B7AE0F-14A7-4882-91BD-ADF2B9DCE5C6} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
+		{D59FC5C7-11E8-4BE6-A500-74088D74F841} = {4B9276C3-651E-48D3-B3A7-3F4D74F3D01A}
 	EndGlobalSection
 EndGlobal
