+++ starsky/starskytest/starsky.foundation.database/QueryTest/InjectServiceScopeTest.cs
@@ -1,7 +1,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using starsky.foundation.database.Query;
 
-namespace starskytest.starsky.foundation.database.Query
+namespace starskytest.starsky.foundation.database.QueryTest
 {
 	[TestClass]
 	public class InjectServiceScopeTest
