+++ starsky/starskytest/starsky.foundation.database/QueryTest/QueryAddRangeTest.cs
@@ -9,7 +9,8 @@ using starsky.foundation.database.Models;
 using starsky.foundation.database.Query;
 using starsky.foundation.platform.Models;
 using starskytest.FakeMocks;
-namespace starskytest.starsky.foundation.database.Query
+
+namespace starskytest.starsky.foundation.database.QueryTest
 {
 	
 	[TestClass]
